make dbCAN easyCGC and easySubstrate work as well

+++ modules/nf-core/rundbcan/easysubstrate/main.nf
@@ -15,7 +15,7 @@ process RUNDBCAN_EASYSUBSTRATE {
     output:
     tuple val(meta), path("${prefix}_overview.tsv")            , emit: cazyme_annotation
     tuple val(meta), path("${prefix}_dbCAN_hmm_results.tsv")   , emit: dbcanhmm_results
-    tuple val(meta), path("${prefix}_dbCANsub_hmm_results.tsv"), emit: dbcansub_results
+    tuple val(meta), path("${prefix}_dbCANsub_hmm_results.tsv"), emit: dbcansub_results, optional: true
     tuple val(meta), path("${prefix}_diamond.out")             , emit: dbcandiamond_results
     tuple val(meta), path("${prefix}_cgc.gff")                 , emit: cgc_gff
     tuple val(meta), path("${prefix}_cgc_standard_out.tsv")    , emit: cgc_standard_out
@@ -46,7 +46,9 @@ process RUNDBCAN_EASYSUBSTRATE {
 
     mv overview.tsv             ${prefix}_overview.tsv
     mv dbCAN_hmm_results.tsv    ${prefix}_dbCAN_hmm_results.tsv
-    mv dbCANsub_hmm_results.tsv ${prefix}_dbCANsub_hmm_results.tsv
+    if [ -f dbCANsub_hmm_results.tsv ]; then
+        mv dbCANsub_hmm_results.tsv ${prefix}_dbCANsub_hmm_results.tsv
+    fi
     mv diamond.out              ${prefix}_diamond.out
     mv cgc.gff                  ${prefix}_cgc.gff
     mv cgc_standard_out.tsv     ${prefix}_cgc_standard_out.tsv
